Cluster Serving docker update (#2681)
