+++ docker/cluster-serving/image_classification_and_object_detection_quick_start.py
