+++ scripts/cluster-serving/cluster-serving-demo.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-source ./cluster-serving-prepare.sh
+source ./cluster-serving-setup.sh
 ./cluster-serving-start -c ./config.yaml & sleep 20
 python3 ./quick_start.py --image_path ../test_image
